init: Adapt sysfs paths for frequency governor nodes.
The reference init script using a for-loop with wildcards
in init.qcom.post_boot.sh to determine the correct sysfs nodes.
We don't have that file, but converted this to a init.qcom.power.rc file.
Normally this nodes shouldn't be changed, but somehow it's happened...

adb shell find /sys/devices -type d -name "qcom,mincpubw*" \
-o -name "qcom,cpubw*" \
-o -name "qcom,gpubw*" \
-o -name "qcom,bcl*" \
| grep -v devfreq

/sys/devices/soc.0/qcom,gpubw.38
/sys/devices/soc.0/qcom,bcl.57
/sys/devices/soc.0/qcom,mincpubw.51
/sys/devices/soc.0/qcom,cpubw.50

Change-Id: I50594c52e931e5edaa4dab7143427376fb20a808